requirements.txt
requests 
pyyaml
